Use React.createClass to avoid problems with this
Apparently something has changed in a recent version of react-onclickoutside
library, causing "this" to be undefined in handleClickOutside. As a result, the
"click outside" functionality broke. When using ES6 classes, "this" is not
automatically bound to the instance, so that is probably causing issues.
Tried changing to React.createClass syntax, which appears to fix the issue.