Patch the Node module loader to avoid issues with peer dependencies
React et.al is defined as peerDependencies in nrfconnect-core/package.json.
These are installed at development time, but they are not bundled with the
library, so those who depend on nrfconnect-core need to add their own React
dependency in their package.json. This works fine when the library is built and
published, but during development there will be a node_modules directory (with
React) in nrfconnect-core, and this causes problems.

When other projects import React components from nrfconnect-core, the Node
module loader will first look for React in nrfconnect-core/node_modules. The
other projects will provide their own React dependency, and that leads to two
copies of React being loaded at the same time. In general, we do not want
anything from nrfconnect-core/node_modules to be loaded by other projects. To
avoid this we patch Node's module loader.

For more info see:
https://www.sharpoblunto.com/News/2016/01/25/testing-react-component-libraries-using-npm-link