Add DFU feature for devices with Secure DFU service
Peripheral devices that have the Secure DFU service will now have a
DFU button. When clicking this button, a dialog appears where the
user selects a zip file to be used for DFU. During transfer, a
progress bar and a throughput graph is displayed.

Right before DFU starts, the application is set to ignore events from
the DFU target device. This is because the application and the DFU
module from pc-ble-driver-js share the same adapter instance. We do
not want the application to respond to events at this time, as it
may interfere with the DFU operation.