Fix loading of spinner image
When building nrfconnect-core, webpack bundles the Spinner component and
ajax-loader.gif into nrfconnect-core/dist. In nrfconnect-appmodule-ble,
we import the Spinner component like this:

import { Spinner } from 'nrfconnect-core';

Webpack looks for import statements, and bundles everything that is
imported.  When building nrfconnect-appmodule-ble, webpack sees the
Spinner import and bundles that, but it does not know that it has to
bundle ajax-loader.gif as well.

One way to solve this is to use a webpack plugin as described here:
http://stackoverflow.com/questions/36559850/include-assets-from-webpack-bundled-npm-package

I am not a fan of this approach. It should be possible to use
nrfconnect-core without this extra copy step. Instead, I have added an
image prop to the Spinner component. This way, the application that uses
Spinner is responsible for importing the image, which allows webpack to
properly bundle it.