Skip copying of pc-ble-driver.dll in setup.bat
I have had a look in old build jobs, and it appears that this copying step has
not been working for a while. In all build logs, including the release-1.0.0,
this step prints "The system cannot find the path specified". Removing it, as
the application does not seem to need it.