Ignore attValue changes when state path does not exist
In some cases, f.ex. during DFU, attributeValueChanged() may be
invoked when the user has not expanded the attribute for the device in
the UI. Then there will be no path to the attribute in state, and it
will try to update the attribute value for a path that does not exist.

To avoid this problem, we ignore cases when an attribute state path
cannot be found.