Add --no-prune flag to electron-packager
By default, electron-packager removes everything in node_modules that
are not specified in the 'dependencies' list in
nrfconnect-loader/package.json. Our own packages are not defined there
anymore (moved to 'links'), so we have to disable the pruning so that
they are included in the final artifact.

This will of course mean that the 'devDependencies' in nrfconnect-loader
are also included, even though they are not required. These dependencies
do not bring a lot of overhead, so keeping it simple for the time being.