Skip unnecessary transpilation
We have been transpiling all our code down to ES5, even though Chrome
v51 that comes with Electron 1.2.8 is fully compatible with ES2015:
http://kangax.github.io/compat-table/es6/#chrome51

Removing transpilation that we do not need by ditching
babel-preset-es2015. We still need babel-preset-react of course, so
that we can use JSX. Also, webpack v1 does not support the
import/export keywords, so we must transpile ES2015 modules to commonjs
modules. When upgrading to webpack v2 (still in beta), we can get rid
of this as well.

There are also a couple of other transforms that are required to build
the project: Default module exports, class properties, and object rest
spread. Not entirely sure why we need the last two, but the build
fails if they are not included, so keeping them for now.

Webpack's UglifyJsPlugin is only able to minify ES5 code, so removing
the minification step. It has minimal impact on load-time anyway:
https://discuss.atom.io/t/minification-of-js-css-for-atom-shell/15528