Upgrade to jest 17.0.2
With recent versions of jest we can avoid invoking webpack. Instead, the
recommended approach is now to just mock the modules that jest cannot parse, by
using moduleNameMapper, as mentioned here:
https://github.com/istarkov/babel-plugin-webpack-loaders

The babel-plugin-webpack-loaders package pulls in a lot of babel dependencies,
so it is best to avoid it if we can.