Mock native modules when testing
Native modules is causing some problems when testing on build servers.
The modules are built to work with Electron v1.2.8, but when running
tests, the node binary on the system is used. If the V8 version on
the system differs from Electron's V8 version, the tests may fail with
'module version mismatch' errors.

In any case, native modules should not have to be loaded when doing
unit tests. They are not really needed, and it will slow down the
tests. Mocking them to avoid problems.