Send firmware content instead of paths to pc-nrfjprog-js
The pc-nrfjprog-js library has been updated to support passing firmware
data as strings. This makes it easier to use from nRF Connect, as we
can just use webpack to load the firmware files into the bundle as
strings, and pass those to pc-nrfjprog-js.

An alternative would be to bundle the firmware files as hex files in
dist/, but then we have to deal with filesystem paths, and that can be
prone to errors.