Fix overflow issue with log viewer
For some reason, the log viewer did not overflow correctly. If a log
entry was very long, the width of the log viewer expanded to fit the
length of the log entry. This messed up the layout, and pushed the
discovered devices section out of the viewport. In version 1.0, the
log viewer clipped the log entries and displayed a scroll bar.

Using position relative/absolute seems to fix the issue.