Include hex files for 1M baud rate
1M is currently not supported on OSX, so we still have to use 115k2
there.

As more hex files are being bundled together with the app, the main
bundle.js file is growing a bit large. This file is loaded when the
application starts, and we do not want to increase initial load time.

Using webpack code splitting to create a separate bundle for hex files,
so that firmware data can be loaded on demand.