Fix issue with empty devices appearing in discovered devices list
The discoveryReducer listens to connect/disconnect actions, and
updates its list of devices accordingly. Sometimes a connect or
disconnect action can be dispatched for a device that is not in
the discovered devices list. Happens f.ex. if the list has been
cleared while a connection exists, and then disconnecting.

Now only updating state if the device exists. Also added tests to
verify that the fix works, plus some extra tests to improve test
coverage of this reducer.