Move LICENSE file for the application to nrfconnect-loader
The LICENSE file that we bundle with the release artifact should be
different from the LICENSE in the root of the project (shown on
GitHub). The nrfconnect-loader package is what we are building and
releasing, so it makes sense to put the application LICENSE there.

Updating setup.bat and setup.sh so that LICENSE is copied from the
right location.