Fix incorrect display of input values in server setup
When switching between service/characteristic/descriptor items, I
noticed that the checkbox values did not update correctly. When
viewing a characteristic with f.ex. "write" checked and then
switching to a characteristic that did not have write enabled, the
checkbox did not update to unchecked.

React has this concept of controlled vs uncontrolled input:
https://facebook.github.io/react/docs/forms.html

When passing undefined or null as value to an input component,
React will assume that the component is uncontrolled, and not
update the input value when rendering. We were passing undefined to
the checkboxes to indicate unchecked state, and this caused the
checkboxes to keep their previous state. It seems that this behavior
was introduced/changed in React 15.

Now making sure that undefined is not passed to the input components,
so that we avoid this issue.

I also see that we are keeping values in instance variables
(this.*) and using forceUpdate() to trigger re-render. Force update
is not recommended, so we might want to reconsider that approach:
https://facebook.github.io/react/docs/react-component.html#forceupdate