Install VC 2015 Redistributable as part of installer
We have been copying Visual C++ Redistributable DLLs into the
pc-ble-driver-js build directory in node_modules. With the new version
of Electron it seems that these are not loaded anymore. Because of
this, users that do not have VC C++ redist 2015 x86 installed will see
a blank/white screen when starting nRF Connect.

Instead of copying these files, we can bundle the redistributable
installer from Microsoft and run it as part of the nRF Connect
installation.