Remove uuid_definitions.json in root of repository
The uuid_definitions.json in the root of the repo is the same as
lib/utils/custom_definitions.json. The custom_definitions.json file is
copied into userDataDir if it does not exist. The uuid_definitions.json
on root does, however, not seem to be used anywhere, so removing it.