Only set libs available at runtime as externals in webpack config
We cannot just use the dependencies from pc-nrfconnect-devdep as
externals. The devdep project has f.ex. react-bootstrap as a
dependency. Having this as an external during development works
fine, as then it is available in node_modules. However, when
publishing the app it will be gone. Only the libs that explicitly
made available to the app (ref. lib/windows/app/index.js in core)
will be available at runtime in production.

Changing the webpack config so that it only considers the libs
that are available during runtime as externals. Also adding
dependencies from the app's package.json as externals, as those
are always installed when installing the app, and do not need to
be bundled by webpack.