Fix nav keys for DeviceDetails and ServerSetup
After the app rewrite, it was no longer possible to use the
arrow keys to navigate up/down and expand items in
DeviceDetails and ServerSetup. Turns out that there is a
(pretty hefty) generator function in treeViewKeyNavigation.js
and that has to be invoked using `for..of`.

Also now using withHotkey instead of the old way of creating
key bindings.