Use UUID constants instead of looking up by name
Instead of using getUuidByName() to lookup UUIDs, we now refer
to UUID constants. The way we were using getUuidByName() was
problematic because custom UUIDs defined by the user were also
included in the search. E.g. if the user had defined a custom
"Secure DFU" UUID, the DFU button would not be displayed for
a connected device even though the device had the Secure DFU
service.