Add custom firmware dialog text
By default, the firmware dialog asks "Would you like to program the
development kit on ... with the required firmware?". For the BLE app we
know that we will always require a connectivity firmware. Also, the
dialog will only be shown if there is no connectivity firmware or if the
connectivity firmware is not the latest.