Log AdapterFactory errors instead of showing dialog
The AdapterFactory checks for new adapters every 2 seconds. If something
fails in this routine, error may be emitted multiple times. Showing an
error dialog in these cases will make the completely app unusable, so
logging the error it instead. It will show up in the log viewer and in
the log file.