Keep connection open when performing DFU
When DFU has started, we do not want the connected device or its events
to be visible in nRF Connect anymore. nRF Connect will respond to events
and interfere with the DFU operation.

To solve this, we have disconnected from the device before handing it
over to the DFU module in pc-ble-driver-js, which then reconnects before
performing DFU. However, some devices may have stopped advertising at
this point, making pc-ble-driver-js unable to connect.

Now hiding the device in the UI, but keeping the connection open so that
pc-ble-driver-js does not have to reconnect.