Change BLE driver event interval from 10 ms to 0
When opening an adapter with pc-ble-driver-js, it is possible to specify
an `eventInterval` option, which is how often BLE events should be sent
up from pc-ble-driver to pc-ble-driver-js. This has been set to 10 ms to
avoid flooding the UI with events, which could make it unresponsive.

However, many changes has been made to nRF Connect since the event
interval mechanism was introduced, and recent tests indicate that
`eventInterval: 0` (which disables the event inverval) does not cause
problems for the UI.

The main motivation for disabling the event interval is to avoid slowing
down the DFU. The DFU procedure waits for events when performing write
operations, and we want these events to arrive as fast as possible.