Fix issue when resetting and applying server setup
The CentralDevice and ServerSetup components require a `deviceInfo`
prop. This device info needs to be loaded from nrfjprog and added to
state before opening the adapter. If not, the CentralDevice and
ServerSetup will fail to render.

Added a short-term fix for this in `resetAndApplyServer` now, but I
think this needs to be refactored. We should not have to remember to
dispatch LOCAL_DEVICE_INFO_LOADED before opening the adapter. This is
easy to forget, and will lead to bugs like this. If it is required for
opening the adapter, then it should probably be handled by
`AdapterActions.openAdapter` instead.