Hide menu items if their related callbacks do not exist
When no default props are defined for a component, they are set as
`undefined` by default. A while ago, default props were added for this
component, and the on* props were set to `null` by default, so checking
for `undefined` here does not work anymore. Fixing this so that the
menu items are hidden when their corresponding on* props are not given.