Fix creating characteristics in the server setup
Creating and saving a characteristic crashed the whole app.

This was caused by a subtle change in immutable: When merging in
properties that contain an array immutable before converted them to
instances of List. Now it keeps them as arrays. But for value we
depended elsewhere on it being a List.